Puzzles (with bugs)
Added puzzles to each room: added radio to first room, padlock and keypad to second room, padlock to third room
radio bug: can only inspect once? after opening with tab, closing with Escape, can't reopen with tab
padlock bug: doesn't have inspect/interactable attached, so the puzzle is just running inside the normal scene. that is, moving around with WASD triggers the rotation of the padlock dials. actually kinda cool idea but we should fix this lol
keypad: literally just a squished cube at this point, renamed in scene to remind whoever implements it that the setup is in the youtube tutorial

Aside from that:
- Fixed Player and HUD instances in each level to work with interactable scripts
- Added 3D audio to the radio so that we can lure players to it by sound alone